refactor: ui의 props가 되는 인터페이스를 공통 인터페이스로 분리합니다.
